Somewhat working state

+++ Assets/Scripts/Base/Runtime/BaseSaveSystem/SaveSystemEditor.cs
@@ -11,41 +11,33 @@ using Sirenix.OdinInspector.Editor;
 using Sirenix.Utilities;
 using Sirenix.Utilities.Editor;
 #endif
-namespace Base
-{
-    public class SaveSystemEditor
-    {
+namespace Base {
+    public class SaveSystemEditor {
         List<SaveObject> Objects;
         Dictionary<string, SaveObject> savesDic;
-        public Task SaveSystemStrapping()
-        {
+        public Task SaveSystemStrapping() {
             Objects = new List<SaveObject>();
             Objects = Resources.LoadAll<SaveObject>("SaveAssets").ToList();
             return Task.CompletedTask;
         }
 
-        public SaveObject GetSaveObject(object obj)
-        {
+        public SaveObject GetSaveObject(object obj) {
             if (savesDic.ContainsKey(obj.ToString())) { return savesDic[obj.ToString()]; }
             return null;
         }
 
-        public SaveSystemEditor()
-        {
+        public SaveSystemEditor() {
             Objects = new List<SaveObject>();
             Objects = Resources.LoadAll<SaveObject>("SaveAssets").ToList();
             savesDic = new Dictionary<string, SaveObject>();
-            for (int i = 0; i < Objects.Count; i++)
-            {
+            for (int i = 0; i < Objects.Count; i++) {
                 Objects[i].LoadThisData();
                 savesDic.Add(Objects[i].SaveName, Objects[i]);
             }
         }
 
-        public void SaveAllData()
-        {
-            foreach (var item in Objects)
-            {
+        public void SaveAllData() {
+            foreach (var item in Objects) {
                 item.SaveThisData();
             }
         }
@@ -53,13 +45,11 @@ namespace Base
 
         #region Editor Functions
 #if UNITY_EDITOR
-        public SaveSystemEditor(OdinMenuTree tree)
-        {
+        public SaveSystemEditor(OdinMenuTree tree) {
             Objects = new List<SaveObject>();
             Objects = Resources.LoadAll<SaveObject>("SaveAssets").ToList();
             savesDic = new Dictionary<string, SaveObject>();
-            for (int i = 0; i < Objects.Count; i++)
-            {
+            for (int i = 0; i < Objects.Count; i++) {
                 savesDic.Add(Objects[i].SaveName, Objects[i]);
             }
             tree.AddAllAssetsAtPath("Saves", "Assets/Resources/SaveAssets");
@@ -68,8 +58,7 @@ namespace Base
         [HorizontalGroup("Split")]
         [VerticalGroup("Split/Right")]
         [Button]
-        public void CreateNewSave()
-        {
+        public void CreateNewSave() {
             SaveObject obj = ScriptableObject.CreateInstance<SaveObject>();
             obj.name = "Save_" + (Objects.Count + 1);
             obj.SaveName = obj.name;
@@ -79,25 +68,25 @@ namespace Base
         [HorizontalGroup("Split")]
         [VerticalGroup("Split/Left")]
         [Button]
-        public void ClearAllSaves()
-        {
+        public void ClearAllSaves() {
             string[] assetsPath = { "Assets/Resources/SaveAssets" };
-            foreach (var item in AssetDatabase.FindAssets("", assetsPath))
-            {
+            foreach (var item in AssetDatabase.FindAssets("", assetsPath)) {
+
                 var path = AssetDatabase.GUIDToAssetPath(item);
-                AssetDatabase.DeleteAsset(path);
+                SaveObject var = AssetDatabase.LoadAssetAtPath(path, typeof(SaveObject)) as SaveObject;
+                if (!var.IsPermanent)
+                    AssetDatabase.DeleteAsset(path);
             }
             AssetDatabase.SaveAssets();
         }
         [HorizontalGroup("Split")]
         [VerticalGroup("Split/Left")]
         [Button]
-        public void CreateEnums()
-        {
+        public void CreateEnums() {
             string[] _temp = new string[Objects.Count];
-            for (int i = 0; i < Objects.Count; i++)
-            {
+            for (int i = 0; i < Objects.Count; i++) {
                 _temp[i] = Objects[i].name;
+                Objects[i].CreateEnums();
             }
             EnumCreator.CreateEnum("Saves", _temp);
         }
